AStarNode uint16_t
git-svn-id: https://felix1991.mine.nu/svn/Earth2150@39 92bb1c38-9b61-c64a-b0c3-bf70b591afdc

+++ cpp/earth2150/include/e2150/AStarNode.h
@@ -6,21 +6,21 @@
 
 class AStarNode{
 	private:
-		static const uint32_t MULTIPLICATOR;
-		static const uint32_t SQRT_1;
-		static const uint32_t SQRT_2;
+		static const uint16_t MULTIPLICATOR;
+		static const uint16_t SQRT_1;
+		static const uint16_t SQRT_2;
 
 		AStarNode *previousNode;
 		uint32_t overallCost;
 		uint32_t spentCost;
 		uint32_t estimationCost;
-		uint32_t x;
-		uint32_t y;
+		uint16_t x;
+		uint16_t y;
 	public:
-		static uint32_t farDistance(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1);
-		static uint32_t nearDistance(uint32_t x0, uint32_t y0, uint32_t x1, uint32_t y1){return (x0 == x1 || y0 == y1) ? SQRT_1 : SQRT_2;}
+		static uint32_t farDistance(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
+		static uint32_t nearDistance(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1){return (x0 == x1 || y0 == y1) ? SQRT_1 : SQRT_2;}
 
-		AStarNode(AStarNode *previousNode, uint32_t spentCost, uint32_t estimationCost, uint32_t x, uint32_t y);
+		AStarNode(AStarNode *previousNode, uint32_t spentCost, uint32_t estimationCost, uint16_t x, uint16_t y);
 
 		uint32_t getEstimationCost() {
 			return estimationCost;
@@ -30,8 +30,8 @@ class AStarNode{
 		MapPosition getPosition() const{return MapPosition(x, y);}
 		AStarNode *getPreviousNode() const{return previousNode;}
 		uint32_t getSpentCost() const{return spentCost;}
-		uint32_t getX() const{return x;}
-		uint32_t getY() const{return y;}
+		uint16_t getX() const{return x;}
+		uint16_t getY() const{return y;}
 };
 
 #endif // ASTARNODE_H
